requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.24.3

# Database
sqlite3-python==1.0.0

# API clients (LLM integration)
google-generativeai==0.3.0
openai==1.3.0

# Data validation & processing
pydantic==2.5.0
faker==20.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & monitoring
python-json-logger==2.0.7

# Development (optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0
pylint==3.0.3
